CLI Menu

+++ dev/src/Business/ShiftPKG/Shift.java
@@ -4,11 +4,12 @@ import Business.Type.RoleType;
 import Business.Type.ShiftType;
 import org.apache.log4j.Logger;
 
+import javax.swing.*;
 import java.time.LocalDate;
 import java.util.*;
 
 public class Shift {
-    final static Logger log = Logger.getLogger(Shift.class);
+    private final static Logger log = Logger.getLogger(Shift.class);
     //-------------------------------------fields------------------------------------
     private final int SID;
     private Map<Integer, String[]> employees;  // [0]->RoleType in String, [1]->name
@@ -18,13 +19,14 @@ public class Shift {
     private LocalDate date;
     private ShiftType shiftType;
     private boolean wasSelfMake;
+    private boolean hasShiftManager;
 
     //------------------------------------constructor--------------------------------
-    public Shift(int SID,Map<RoleType, Integer> rolesAmount, Map<RoleType, List<String[]>> optionals, LocalDate date, ShiftType shiftType) throws Exception {
+    public Shift(int SID, Map<RoleType, Integer> rolesAmount, Map<RoleType, List<String[]>> optionals, LocalDate date, ShiftType shiftType) throws Exception {
 
         this.SID = SID;
         employees = new HashMap<>();
-        checkIfAmountNegative(rolesAmount);  //check if legal (not negative)
+        checkIfAmountNegAndHasSM(rolesAmount);  //check if legal (not negative)
         this.rolesAmount = rolesAmount;
         this.optionals = optionals;
         this.complete = false;
@@ -36,39 +38,64 @@ public class Shift {
         this.shiftType = shiftType;
         this.wasSelfMake = false;
         log.debug("shift: " + SID + " created");
+        if(optionals.containsKey(RoleType.ShiftManager))
+            hasShiftManager = !optionals.get(RoleType.ShiftManager).isEmpty();
+        ;
+    }
+
+    public boolean HasShiftManager() {
+        return hasShiftManager;
+    }
+
+    public List<Integer> insertShiftManager() {
+        String[] sm = optionals.get(RoleType.ShiftManager).remove(0);
+        employees.put(Integer.parseInt(sm[0]), new String[]{RoleType.ShiftManager.name(), sm[1]});
+        List<Integer> listOfSM = new LinkedList<>();
+        listOfSM.add(Integer.parseInt(sm[0]));
+        return listOfSM;
     }
 
     //-------------------------------------methods----------------------------------
 
     public List<Integer> self_make() throws Exception {
-        boolean comp = true;
-        for (Map.Entry<RoleType, Integer> e : rolesAmount.entrySet()) {
-            RoleType role = e.getKey();
-            int amount = e.getValue();
-            while (amount > 0) {
-                if (optionals.get(role)==null || optionals.get(role).isEmpty()) {
-                    comp = false;
-                    break;
-                } else {
-                    String[] emp = optionals.get(role).remove(0);  // delete from optionals
-                    employees.put(Integer.parseInt(emp[0]), new String[]{role.name(), emp[1]});  //add to employees
-                    RemoveEmpFromOptionals(Integer.parseInt(emp[0]));
-                    amount--;
+        if (!wasSelfMake) {
+            for (Map.Entry<RoleType, Integer> e : rolesAmount.entrySet()) {
+                if (e.getKey().equals(RoleType.ShiftManager)) continue;
+                RoleType role = e.getKey();
+                int amount = e.getValue();
+                while (amount > 0) {
+                    if (optionals.get(role) == null || optionals.get(role).isEmpty()) {
+                        break;
+                    } else {
+                        String[] emp = optionals.get(role).remove(0);  // delete from optionals
+                        employees.put(Integer.parseInt(emp[0]), new String[]{role.name(), emp[1]});  //add to employees
+                        RemoveEmpFromOptionals(Integer.parseInt(emp[0]));
+                        amount--;
+                    }
                 }
             }
+            complete = isComplete();
+            wasSelfMake = true;
+            return new ArrayList<>(employees.keySet());
         }
-        complete = comp;
-        wasSelfMake = true;
-        return new ArrayList<>(employees.keySet());
+        return new ArrayList<>();
     }
 
-    public void removeEmpFromShift(int EID) throws Exception {
+
+    public void removeEmpFromShift(int EID, List<RoleType> roles) throws Exception {
         String[] emp = employees.remove(EID);  //return null if EID isn't in this shift
         if (emp == null) {
             log.error("EID: " + EID + " isn't in this shift");
             throw new Exception("EID: " + EID + " isn't in this shift");
         }
         complete = false;
+        if (RoleType.valueOf(emp[0]).equals(RoleType.ShiftManager)) {
+            hasShiftManager = false;
+        }
+        roles.forEach(roleType -> {
+            optionals.get(roleType).add(new String[]{String.valueOf(EID), emp[1]});
+        });
+
         log.debug("EID: " + EID + "removed from SID: " + SID);
     }
 
@@ -76,7 +103,7 @@ public class Shift {
         //check if EID can work in this shift (in optionals)
         List<String[]> list = optionals.get(role);
         boolean canWork = false;
-        if(list!=null) {
+        if (list != null) {
             for (String[] s : list) {
                 if (Integer.parseInt(s[0]) == EID && s[1].equals(name)) {
                     canWork = true;
@@ -86,7 +113,7 @@ public class Shift {
         }
         if (!canWork) {
             log.error("EID: " + EID + " isn't option to be: " + role + " in SID: " + SID);
-            throw new Exception("EID: " + EID + " isn't option to be " + role + " in this shift");
+            throw new Exception("Employee ID: " + EID + " isn't option to be " + role + " in this shift");
         }
         //check if there is empty role for this roleType in rolesAmount
         if (roleIsFull(role)) {
@@ -96,6 +123,8 @@ public class Shift {
         }
         employees.put(EID, new String[]{role.name(), name});
         RemoveEmpFromOptionals(EID);
+        complete = isComplete();
+        hasShiftManager = role.equals(RoleType.ShiftManager);
         log.debug("EID: " + EID + "added to SID: " + SID);
     }
 
@@ -126,43 +155,61 @@ public class Shift {
         }
     }
 
-    public String[] addToOptionals(int EID, String name, List<RoleType> role) {
+    public String[] addToOptionals(int EID, String name, RoleType role) {
         String[] s = {String.valueOf(EID), name};
-        for (RoleType roleType : role) {
-            List<String[]> l = optionals.get(roleType);
-            if(l==null){
-                ArrayList<String[]> list = new ArrayList<>();
-                list.add(s);
-                optionals.put(roleType,list);
-            }else{
-                l.add(s);
-            }
-        }
+        optionals.get(role).add(s);
         return s;
     }
 
     public void removeFireEmp(int EID, String name) {
         String[] s = employees.remove(EID);  //remove employee if he is in this shift
-        if (s != null)
+        if (s != null) {
             complete = false;
+            if (RoleType.valueOf(s[0]).equals(RoleType.ShiftManager)) {
+                hasShiftManager = false;
+            }
+        }
         for (Map.Entry<RoleType, List<String[]>> o : optionals.entrySet()) {
             List<String[]> listOfEmp = o.getValue();
-            if(listOfEmp!=null) {
+            if (listOfEmp != null) {
                 String[] toRemove = {String.valueOf(EID), name};
                 listOfEmp.remove(toRemove);
             }
         }
-        log.debug("EID: "+EID + " removed(fire) from SID: "+ SID);
+        log.debug("EID: " + EID + " removed(fire) from SID: " + SID);
     }
 
-    private void checkIfAmountNegative(Map<RoleType, Integer> defaultShifts) throws Exception {
+    private void checkIfAmountNegAndHasSM(Map<RoleType, Integer> defaultShifts) throws Exception {
         for (Map.Entry<RoleType, Integer> e : defaultShifts.entrySet()) {
             if (e.getValue() < 0) {
-                log.error("Role amount for role: "+ e.getKey() + " is negative: " +e.getValue());
+                log.error("Role amount for role: " + e.getKey() + " is negative: " + e.getValue());
                 throw new Exception("Role amount for role: " + e.getKey() + " is negative");
+            }
+            if (e.getKey().equals(RoleType.ShiftManager) && e.getValue()!=1) {
+                log.error("Shift Manager amount is not legal: "+ e.getValue());
+                throw new Exception("Shift Manager amount is not legal: "+ e.getValue());
             }
 
         }
+    }
+
+    public void updateName(int eid, List<RoleType> roles, String newName) {
+        if(employees.containsKey(eid))
+            employees.get(eid)[1] = newName;
+        roles.forEach(roleType -> {
+            List<String[]> emps = optionals.get(roleType);
+            emps.forEach(pair -> {
+                if(Integer.parseInt(pair[0]) == eid)
+                    pair[1] = newName;
+            });
+        });
+    }
+
+    private boolean isComplete() {
+        int amount = 0;
+        for (Map.Entry<RoleType, Integer> role : rolesAmount.entrySet())
+            amount += role.getValue();
+        return amount == employees.size();
     }
 
     //-------------------getters&setters----------------------------------------
